feat(inventory): FIFO cost layers + Opening Stock module
Stages (b) and (c) of the FIFO inventory costing ticket — see
prompt-fifo-page.md. Stage (a) investigation was presented in chat first,
per the ticket's own request to investigate before writing code.

Stage (b) — cost layer schema + core logic:
- New fifo_layers (one row per receiving event) and fifo_layer_consumptions
  (audit trail, one row per layer touched per consuming voucher — makes
  COGS-per-voucher and exact-layer reversal possible without recomputing
  anything) tables.
- FifoLayerService: receive()/consume() (oldest-first, rejects outright on
  insufficient stock rather than a zero-cost layer), reverseReceipt()/
  reverseConsumption() (restores the *exact* original layers, not a new
  averaged-cost one), and averageConsumedCost() for the two cases that
  can't do an exact restore (Stock Transfer's destination leg, Credit Note
  restock).
- Wired into all 7 real write paths: Goods Receipt, Delivery, Stock
  Transfer, Stock Adjustment (new required Unit Cost input for a
  found-more-than-expected line), Purchase Return (submit=consume,
  reverse=exact restore), and Credit Note restock — previously "intent
  only, never wired" per its own docblock, now a real stock movement
  exactly as docs/CREDIT_NOTE_DESIGN.md §5 already specified.
- Fixed two pre-existing decimal-qty gaps found during investigation:
  DeliveryItem.qty and CreditNoteItem.qty_credited were integer-only
  unlike every sibling line-item table.
- 22 existing test files seeded stock by calling StockLedgerService
  directly (bypassing Goods Receipt), which broke once Delivery/Transfer/
  etc. started requiring a real FIFO layer to consume — fixed once via a
  shared TestCase::seedStock() helper instead of patching each file's
  business logic.

Stage (c) — Opening Stock module:
- New OpeningStock/OpeningStockItem documents (standard Draft->Submitted->
  Cancelled lifecycle) — the one stock document in this app where cancel()
  actually works, since FifoLayerService::reverseReceipt() already blocks
  it once a layer is partially consumed.
- Bulk CSV/XLSX import reuses the existing 1-step import engine, but each
  row is a whole new document (header + one line item) rather than the
  generic single-flat-record shape every other import target has — added
  a small opt-in CreatesRelatedRecords interface so ProcessImportBatchJob
  can delegate to OpeningStockImportTemplate::persist() (which itself just
  calls OpeningStockService::create(), reusing all the same validation the
  manual page runs) instead of its default $model::updateOrCreate().
- Every imported document is tagged with its ImportBatch id; the list page
  can filter to one batch and Submit All/Cancel All it in one transaction
  (Cancel All pre-checks every document before cancelling any, naming
  every already-consumed one that blocks it).

+++ backend/laravel-api/app/Http/Requests/UpdateCreditNoteRequest.php
@@ -23,7 +23,7 @@ class UpdateCreditNoteRequest extends FormRequest
             'remarks' => ['nullable', 'string'],
             'items' => ['sometimes', 'array'],
             'items.*.invoice_item_id' => ['required_with:items', 'uuid', 'exists:invoice_items,id'],
-            'items.*.qty_credited' => ['sometimes', 'integer', 'min:0'],
+            'items.*.qty_credited' => ['sometimes', 'numeric', 'min:0'],
             'items.*.amount' => ['required_with:items', 'numeric', 'min:0'],
             'items.*.restock' => ['sometimes', 'boolean'],
         ];
